Refactors the code
